feat: refactor to skills-based architecture with team ownership
Restructure the monolithic common_agents/ directory into 4 self-contained
skills under skills/, each owned by a different team. This enables
independent development and maintenance per domain while sharing a
stable contract layer via shared_libraries/ and session state keys.

Skills created:
- trend-discovery (Data/Analytics Team): trend APIs, BigQuery, YouTube
- market-research (Research/Content Team): parallel research pipeline
- ad-creative (Creative Team): ad copy, visual concepts, Imagen/Veo
- av-studio (AV Production Team): clip chaining, ffmpeg, commercials

Each skill includes SKILL.md metadata, reference docs, and documented
session state contracts. Added SKILLS_GUIDE.md for team onboarding and
3 GCP-branded architecture diagrams.
