+++ trends_and_insights_agent/__init__.py
