--- trends_and_insights_agent/common_agents/ad_content_generator/__init__.py
@@ -1,3 +1,0 @@
-from .agent import ad_content_generator_agent
-
-__all__ = ["ad_content_generator_agent"]
