+++ trends_and_insights_agent/skills/market_research/__init__.py
@@ -1,0 +1,3 @@
+from .agents import research_orchestrator
+
+__all__ = ["research_orchestrator"]
