+++ trends_and_insights_agent/skills/market_research/prompts.py
@@ -1,0 +1,6 @@
+"""Prompts for the market research pipeline agents.
+
+Inline prompts are defined directly in agents.py for this skill since
+they are tightly coupled to the agent definitions. This module is reserved
+for any future extracted or shared prompts.
+"""
