+++ trends_and_insights_agent/skills/market_research/tools.py
