+++ trends_and_insights_agent/skills/trend_discovery/__init__.py
@@ -1,0 +1,3 @@
+from .agents import trends_and_insights_agent
+
+__all__ = ["trends_and_insights_agent"]
