+++ trends_and_insights_agent/skills/trend_discovery/agents.py
